* fixed warnings

+++ src/tools/gtkradiant/1.5/radiant/radiant/qgl.h
@@ -28,6 +28,6 @@ void QGL_sharedContextDestroyed(OpenGLBinding& table);
 
 bool QGL_ExtensionSupported(const char* extension);
 
-float QGL_maxTextureAnisotropy();
+int QGL_maxTextureAnisotropy();
 
 #endif
